% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ce11_ATACseq}
\alias{ce11_ATACseq}
\title{ce11_ATACseq}
\format{
RleList
}
\source{
\href{https://doi.org/10.1101/2020.02.20.958579}{BiorXiv}
}
\usage{
data(ce11_ATACseq)
}
\description{
Sample of ATAC-seq from mixed tissues in C. elegans young adults
}
\examples{
data(ce11_ATACseq)
ce11_ATACseq
}
\references{
Serizay et al. 2020, 
"Tissue-specific profiling reveals distinctive regulatory 
architectures for ubiquitous, germline and somatic genes", BiorXiv.
(\href{https://doi.org/10.1101/2020.02.20.958579}{DOI})
}
\keyword{datasets}
