% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ce11_TSSs}
\alias{ce11_TSSs}
\title{ce11_TSSs}
\format{
GRanges
}
\source{
\href{https://doi.org/10.1101/2020.02.20.958579}{BiorXiv}
}
\usage{
data(proms)
}
\description{
Annotated TSSs in C. elegans (ce11) according to Serizay et 
al. 2020, "Tissue-specific profiling reveals distinctive regulatory 
architectures for ubiquitous, germline and somatic genes", BiorXiv.
}
\examples{
data(ce11_TSSs)
lengths(ce11_TSSs)
}
\references{
Serizay et al. 2020, 
"Tissue-specific profiling reveals distinctive regulatory 
architectures for ubiquitous, germline and somatic genes", BiorXiv.
(\href{https://doi.org/10.1101/2020.02.20.958579}{DOI})
}
\keyword{datasets}
