% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ce11_WW_10bp}
\alias{ce11_WW_10bp}
\title{ce11_WW_10bp}
\format{
RleList
}
\source{
\href{https://doi.org/10.1101/2020.02.20.958579}{BiorXiv}
}
\usage{
data(ce11_WW_10bp)
}
\description{
Sample of WW 10-bp periodicity track generated by getPeriodicityTrack()
in ce11 over annotated accessible sites, with default parameters
}
\examples{
data(ce11_WW_10bp)
ce11_WW_10bp
}
\references{
Serizay et al. 2020, 
"Tissue-specific profiling reveals distinctive regulatory 
architectures for ubiquitous, germline and somatic genes", BiorXiv.
(\href{https://doi.org/10.1101/2020.02.20.958579}{DOI})
}
\keyword{datasets}
