% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ce_proms_seqs}
\alias{ce_proms_seqs}
\title{ce_proms_seqs}
\format{
DNAStringSet
}
\source{
\href{https://doi.org/10.1101/2020.02.20.958579}{BiorXiv}
}
\usage{
data(ce_proms_seqs)
}
\description{
Sample of sequences of promoters annotated in C. elegans (ce11) according to
Serizay et al. 2020, "Tissue-specific profiling reveals distinctive 
regulatory architectures for ubiquitous, germline and somatic genes", 
BiorXiv.
}
\examples{
data(ce_proms_seqs)
head(ce_proms_seqs)
}
\references{
Serizay et al. 2020, 
"Tissue-specific profiling reveals distinctive regulatory 
architectures for ubiquitous, germline and somatic genes", BiorXiv.
(\href{https://doi.org/10.1101/2020.02.20.958579}{DOI})
}
\keyword{datasets}
