% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{char2BSgenome}
\alias{char2BSgenome}
\title{A function to get the BSgenome from genome ID}
\usage{
char2BSgenome(ID)
}
\arguments{
\item{ID}{character}
}
\value{
BSgenome object
}
\description{
A function to get the BSgenome from genome ID
}
\examples{
\dontrun{
    genome <- char2BSgenome('ce11')
    genome
}
}
