% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkPeriodFromFourier}
\alias{checkPeriodFromFourier}
\title{checkPeriodFromFourier}
\usage{
checkPeriodFromFourier(len, period)
}
\arguments{
\item{len}{length of the vector used for FFT}

\item{period}{Float, period of interest}
}
\value{
Float period returned by FFT closest to the period of interest
}
\description{
A function to quickly check whether a period of interest is in 
the result of a FFT analysis, based on the length of the vector which 
is used for FFT
}
\examples{
checkPeriodFromFourier(300, 10)
}
