% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{cleanUpDirectory}
\alias{cleanUpDirectory}
\title{Internal function}
\usage{
cleanUpDirectory(logfile)
}
\arguments{
\item{logfile}{character, name of log file}
}
\value{
NULL Clean-up temporary files generated when running 
getPeriodicityTrack.
}
\description{
Internal function
}
