% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DNAStringSet2GRanges}
\alias{DNAStringSet2GRanges}
\title{A function to generate a GRanges from a DNAStringSet}
\usage{
DNAStringSet2GRanges(seqs)
}
\arguments{
\item{seqs}{A DNAStringSet}
}
\value{
A GRanges object
}
\description{
This function takes a DNAStringSet (for instance a set of sequences
obtained from a BSgenome object) and converts it into 
a GRanges object.
}
\examples{
DNAStringSet2GRanges(
    BSgenome.Celegans.UCSC.ce11::BSgenome.Celegans.UCSC.ce11
)
}
