% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{getChunks}
\alias{getChunks}
\title{Internal function}
\usage{
getChunks(genome.partionned.filtered, PROCS)
}
\arguments{
\item{genome.partionned.filtered}{Output from partitionGenome() filtered 
over GRanges of interest.}

\item{PROCS}{An integer Split the workload over several processors 
(default 12).}
}
\value{
list Chunks list of primary bins to measure.
}
\description{
Internal function
}
