% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges.R
\name{getCovMatrix}
\alias{getCovMatrix}
\title{Internal function}
\usage{
getCovMatrix(granges, bw_file, norm = "none", verbose = TRUE)
}
\arguments{
\item{granges}{A GRanges to map a track onto}

\item{bw_file}{a track in RleList format}

\item{norm}{Should the signal be normalized ('none', 'zscore' or 'log2')?}

\item{verbose}{Boolean}
}
\value{
A square numerical matrix of signal values over the GRanges
}
\description{
Internal function
}
