% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPI.R
\name{getFPI.DNAStringSet}
\alias{getFPI.DNAStringSet}
\title{A function to compute fold power increase (FPI)}
\usage{
\method{getFPI}{DNAStringSet}(
  x,
  motif,
  period = 10,
  n_shuffling = 10,
  cores_shuffling = 1,
  cores_computing = 1,
  verbose = 1,
  ...
)
}
\arguments{
\item{x}{DNAStringSet, sequences of interest}

\item{motif}{character, k-mer of interest}

\item{period}{integer, Period of interest}

\item{n_shuffling}{integer, Number of shuffling}

\item{cores_shuffling}{integer, Number of cores used for shuffling}

\item{cores_computing}{integer, split the workload over several processors 
using BiocParallel}

\item{verbose}{integer, Should the function be verbose?}

\item{...}{Additional arguments}
}
\value{
Several metrics including FPI, observed PSD, etc...
}
\description{
This function takes a set of sequences and a k-mer of interest, 
along with a given period. It calculates the FPI, as introduced
by Pich et al., Cell 2018.
}
\details{
The FPI is calculated as the ratio of the observed k-mer PSD 
at the chosen perdiod subtracted by the median k-mer PSD
in shuffled sequences, divided by the median k-mer PSD 
in shuffled sequences. The number of shufflings is specified 
by the n_shuffling argument.
}
\examples{
data(ce11_proms_seqs)
BiocParallel::register(BiocParallel::SnowParam(workers = 1))
fpi <- getFPI(
    ce11_proms_seqs[1:10], 
    motif = 'TT'
)
fpi$FPI
fpi$observed_PSD
fpi$shuffled_PSD
plotFPI(fpi)
}
