% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPI.R
\name{getFPI.GRanges}
\alias{getFPI.GRanges}
\title{A function to compute fold power increase (FPI)}
\usage{
\method{getFPI}{GRanges}(x, genome, ...)
}
\arguments{
\item{x}{GRanges, GRanges of interest}

\item{genome}{Genome ID or BSgenome}

\item{...}{Additional arguments}
}
\value{
Several metrics including FPI, observed PSD, etc...
}
\description{
This function takes a GRanges and a k-mer of interest, 
along with a given period. It calculates the FPI, as introduced
by Pich et al., Cell 2018.
}
\details{
The FPI is calculated as the ratio of the observed k-mer PSD 
at the chosen perdiod subtracted by the median k-mer PSD
in shuffled sequences, divided by the median k-mer PSD 
in shuffled sequences. The number of shufflings is specified 
by the n_shuffling argument.
}
\examples{
data(ce11_TSSs)
fpi <- getFPI(
    ce11_TSSs[['Ubiq.']][1:10], 
    genome = 'BSgenome.Celegans.UCSC.ce11', 
    motif = 'TT', 
    cores_shuffling = 1
)
fpi$FPI
fpi$observed_PSD
fpi$shuffled_PSD
p1 <- plotFPI(fpi)
fpi <- getFPI(
    ce11_TSSs[['Muscle']][1:5], 
    genome = 'BSgenome.Celegans.UCSC.ce11', 
    motif = 'TT', 
    cores_shuffling = 1
)
fpi$FPI
fpi$observed_PSD
fpi$shuffled_PSD
p2 <- plotFPI(fpi)
cowplot::plot_grid(p1, p2)
}
