% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges.R
\name{getMotifMatrix}
\alias{getMotifMatrix}
\title{Core function - NOT WORKING YET}
\usage{
getMotifMatrix(granges, motif,
  seqs = Biostrings::readDNAStringSet("~/shared/sequences/Caenorhabditis_elegans.WBcel235.dna.toplevel.fa"),
  center = FALSE, flank = NULL, stranded = TRUE,
  split.bid.proms = TRUE, verbose = FALSE)
}
\description{
Core function - NOT WORKING YET
}
