% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPI.R
\name{getPeriodsMetrics}
\alias{getPeriodsMetrics}
\title{A function used to compute l2FC, FPIs and associated p-values}
\usage{
getPeriodsMetrics(fpi)
}
\arguments{
\item{fpi}{result of getFPI() function or getPeriodicity() function with
n_shuffling != 0}
}
\value{
A table with observed PSD values, log2 fold-change and 
associated empirical p-values, and FPI for each frequency 
assessed by getPeriodicity().
}
\description{
A function used to compute l2FC, FPIs and associated p-values
}
\examples{
data(ce11_TSSs)
fpi <- getFPI(
    ce11_TSSs[['Ubiq.']][1:10], 
    genome = 'BSgenome.Celegans.UCSC.ce11', 
    motif = 'TT', 
    cores_shuffling = 1
)
getPeriodsMetrics(fpi)$periodicityMetrics
}
