% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPI.R
\name{getSignificantPeriods}
\alias{getSignificantPeriods}
\title{A function used to compute FPIs and associated p-values}
\usage{
getSignificantPeriods(fpi, pval_threshold = 1)
}
\arguments{
\item{fpi}{result of getFPI() function or getPeriodicity() function with
n_shuffling != 0}

\item{pval_threshold}{Float, the significance threshold (e.g. 0.01 or 0.05).
Please keep in mind that empirical p-values cannot be smaller than 
1/(# of permutations + 1).}
}
\value{
A table with observed PSD values, associated empirical p-values and
FPI for each frequency assessed by getPeriodicity().
}
\description{
A function used to compute FPIs and associated p-values
}
\examples{
data(ce11_TSSs)
fpi <- getFPI(
    ce11_TSSs[['Ubiq.']][1:10], 
    genome = 'ce11', 
    motif = 'TT', 
    cores_shuffling = 1
)
getSignificantPeriods(fpi)$significantPeriods
}
