% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na.remove}
\alias{na.remove}
\title{A tidyverse-compatible function to remove NAs from a vector}
\usage{
na.remove(x)
}
\arguments{
\item{x}{vector}
}
\value{
A vector without NAs
}
\description{
A tidyverse-compatible function to remove NAs from a vector
}
