% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na.replace}
\alias{na.replace}
\title{A function to replace NA by something else in a vector}
\usage{
na.replace(x, value)
}
\arguments{
\item{x}{vector}

\item{value}{Replace NA by this variable}
}
\value{
A vector with NA replaced by value
}
\description{
\code{namedListToLongFormat(x)} returns a data.frame in long format, with 
an added 'name' column, containing the names of the input list.
}
