% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{namedListToLongFormat}
\alias{namedListToLongFormat}
\title{A function to easily coerce a named list into a long data.frame}
\usage{
namedListToLongFormat(x)
}
\arguments{
\item{x}{A named list.}
}
\value{
A long data frame
}
\description{
A function to easily coerce a named list into a long data.frame
}
