% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicity.R
\name{normalizeHistogram}
\alias{normalizeHistogram}
\title{Internal function}
\usage{
normalizeHistogram(hist)
}
\arguments{
\item{hist}{Vector a numeric vector}
}
\value{
a normalized vector
}
\description{
Internal function
}
