% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{partitionBin}
\alias{partitionBin}
\title{Internal function}
\usage{
partitionBin(grange, genome, window_sliding_size, window_sliding_bin)
}
\arguments{
\item{grange}{A GRanges object of length 1}

\item{genome}{DNAStringSet, BSgenome or genome ID}

\item{window_sliding_size}{Integer The width of the bins
to split each primary bin in (default 60).}

\item{window_sliding_bin}{Integer The increment between secondary bins
(default 5).}
}
\value{
list A GRanges object
}
\description{
Internal function
}
