% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges.R
\name{plotAggregateCoverage.CompressedRleList}
\alias{plotAggregateCoverage.CompressedRleList}
\title{A function to plot aggregated signals over sets of GRanges}
\usage{
\method{plotAggregateCoverage}{CompressedRleList}(x, granges, ...)
}
\arguments{
\item{x}{a single signal track (CompressedRleList class)}

\item{granges}{a GRanges object or a list of GRanges}

\item{...}{additional parameters}
}
\value{
An aggregate coverage plot.
}
\description{
This function takes one or several RleList genomic tracks (imported
by rtraklayer::import(..., as = 'Rle')) and one or several GRanges 
objects. It computes coverage of the GRanges by the genomic tracks
and returns an aggregate coverage plot.
}
\examples{
data(ce11_ATACseq)
data(ce11_proms)
class(ce11_ATACseq)
p <- plotAggregateCoverage(
    ce11_ATACseq, 
    resize(ce11_proms[1:100], fix = 'center', width = 1000)
)
p
}
