% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges.R
\name{plotAggregateCoverage}
\alias{plotAggregateCoverage}
\title{A function to plot aggregated signals over sets of GRanges}
\usage{
plotAggregateCoverage(x, ...)
}
\arguments{
\item{x}{CompressedRleList or SimpleRleList (in lists or single)}

\item{...}{additional parameters}
}
\value{
An aggregate coverage plot.
}
\description{
This function takes one or several RleList genomic tracks (imported
by rtraklayer::import(..., as = 'Rle')) and one or several GRanges 
objects. It computes coverage of the GRanges by the genomic tracks
and returns an aggregate coverage plot.
}
\examples{
data(ce11_ATACseq)
data(ce11_proms)
p <- plotAggregateCoverage(
    ce11_ATACseq, 
    resize(ce11_proms[1:100], fix = 'center', width = 1000)
)
p
}
