% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges.R
\name{plotAggregateCoverage.SimpleRleList}
\alias{plotAggregateCoverage.SimpleRleList}
\title{Core function}
\usage{
\method{plotAggregateCoverage}{SimpleRleList}(bw, granges,
  colors = c("#991919", "#1232D9", "#3B9B46", "#D99B12", "#7e7e7e",
  "#D912D4", "#9E7FCC", "#B0E797", "#D1B3B3", "#23A4A3", "#000000"),
  xlab = "Center of elements", ylab = "Score", xlim = NULL,
  ylim = NULL, quartiles = c(0.025, 0.975), verbose = FALSE,
  bin = 1, plot_central = TRUE, ...)
}
\arguments{
\item{bw}{a single signal track SimpleRleList class}

\item{granges}{a GRanges object or a list of GRanges}

\item{colors}{a vector of colors}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{xlim}{y axis limits}

\item{ylim}{y axis limits}

\item{quartiles}{Which quantiles to use to determine y scale automatically?}

\item{verbose}{Boolean}

\item{bin}{Integer Width of the window to use to smooth values}

\item{plot_central}{Boolean Draw a vertical line at 0}
}
\value{
A plot
}
\description{
Core function
}
