% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges.R
\name{plotHeatmaps}
\alias{plotHeatmaps}
\title{Core function - NOT WORKING YET}
\usage{
plotHeatmaps(MAT, colorspace = NULL, autoscale = TRUE, s = 0.05,
  zmin = 0, zmax = 10, titles = rep("", length(MAT)), xlab = "",
  ylab = "", main = "", cex.lab = 12, cex.axis = 12,
  cex.legend = 12, pointsize = 12, cex.title = 20, BIN = 1,
  seriate = FALSE, clusters = 0, ...)
}
\description{
Core function - NOT WORKING YET
}
