% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sampleGRanges.BSgenome}
\alias{sampleGRanges.BSgenome}
\title{A function to sample GRanges from BSgenome}
\usage{
\method{sampleGRanges}{BSgenome}(x, ...)
}
\arguments{
\item{x}{a BSgenome object}

\item{...}{Additional parameters}
}
\value{
A GRanges object of length n
}
\description{
A function to sample GRanges from BSgenome
}
\examples{
\dontrun{
    sampleGRanges(
        (BSgenome.Scerevisiae.UCSC.sacCer3::
            BSgenome.Scerevisiae.UCSC.sacCer3), 
        100
    )
}
}
