% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sampleGRanges.character}
\alias{sampleGRanges.character}
\alias{sampleGenome}
\title{A function to sample GRanges from genome IDs}
\usage{
\method{sampleGRanges}{character}(x, ...)

sampleGenome(x, ...)
}
\arguments{
\item{x}{a genome ID ()}

\item{...}{Additional parameters}
}
\value{
A GRanges object of length n
}
\description{
A function to sample GRanges from genome IDs
}
\examples{
\dontrun{
    sampleGRanges(
        'ce11', 
        100
    )
}
}
