% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sampleGRanges.DNAStringSet}
\alias{sampleGRanges.DNAStringSet}
\title{A function to sample GRanges within DNAStringSet}
\usage{
\method{sampleGRanges}{DNAStringSet}(x, ...)
}
\arguments{
\item{x}{a DNAStringSet object}

\item{...}{Additional parameters}
}
\value{
A GRanges object of length n
}
\description{
A function to sample GRanges within DNAStringSet
}
\examples{
\dontrun{
    data(ce_proms)
    sampleGRanges(ce_proms, 100)
}
}
