% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sampleGRanges.GRanges}
\alias{sampleGRanges.GRanges}
\title{A function to sample GRanges within GRanges}
\usage{
\method{sampleGRanges}{GRanges}(x, n = NULL, width = NULL,
  exclude = TRUE)
}
\arguments{
\item{x}{a GRanges object}

\item{n}{Integer number of sampled GRanges}

\item{width}{Integer width of sampled GRanges}

\item{exclude}{Boolean should the original GRanges be excluded?}
}
\value{
A GRanges object of length n
}
\description{
A function to sample GRanges within GRanges
}
\examples{
\dontrun{
    data(ce_proms)
    sampleGRanges(ce_proms)
}
}
