% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sampleGRanges}
\alias{sampleGRanges}
\title{A function to sample GRanges within GRanges}
\usage{
sampleGRanges(granges, n)
}
\arguments{
\item{granges}{a GRanges object}

\item{n}{Integer}
}
\value{
A GRanges object of length n
}
\description{
A function to sample GRanges within GRanges
}
