% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigwigs.R
\name{scaleBigWigs}
\alias{scaleBigWigs}
\title{A function to z-score a bigWig track imported by rtracklayer.}
\usage{
scaleBigWigs(bigWigs)
}
\arguments{
\item{bigWigs}{a RleList, usually obtained by rtracklayer::import()}
}
\value{
a list of scaled Rle (z-score)
}
\description{
A function to z-score a bigWig track imported by rtracklayer.
}
