% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setUpBPPARAM}
\alias{setUpBPPARAM}
\title{setUpBPPARAM}
\usage{
setUpBPPARAM(nproc = 1)
}
\arguments{
\item{nproc}{number of processors}
}
\value{
A BPPARAM object
}
\description{
A function to dynamically select 
MulticoreParam or SnowParam (if Windows)
}
\examples{
BPPARAM <- setUpBPPARAM(1)
}
