% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shuffleSeq}
\alias{shuffleSeq}
\title{A function to quickly shuffle sequence(s)}
\usage{
shuffleSeq(dna, seed = 17)
}
\arguments{
\item{dna}{DNAString or DNAStringSet}

\item{seed}{Integer}
}
\value{
A DNAString or DNAStringSet
}
\description{
A function to quickly shuffle sequence(s)
}
