% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPI.R
\name{significantPeriods}
\alias{significantPeriods}
\title{Internal function}
\usage{
significantPeriods(fpi, threshold = 1e-04)
}
\arguments{
\item{fpi}{result of getFPI() function}

\item{threshold}{Float, p-value used as significance threshold}
}
\value{
periods from the FPI list at which PSD are statistically higher 
than those from shuffled sequences
}
\description{
Internal function
}
