% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{smoothBigWig}
\alias{smoothBigWig}
\title{Internal function}
\usage{
smoothBigWig(bw, k = 20, BPPARAM)
}
\arguments{
\item{bw}{A RleList object}

\item{k}{Integer, smoothing window}

\item{BPPARAM}{BPPARAM}
}
\value{
bw RleList smoothed (for each chromosome) using a window of k.
}
\description{
Internal function
}
