% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges.R
\name{withSeq}
\alias{withSeq}
\title{Core function}
\usage{
withSeq(granges, genome)
}
\arguments{
\item{granges}{A GRanges object}

\item{genome}{DNAStringSet object. Ideally, the sequence of an entire genome, 
obtained for instance by running 
\code{Biostrings::getSeq(BSgenome.Celegans.UCSC.ce11::BSgenome.Celegans.UCSC.ce11)}.}
}
\value{
GRanges with a seq column containing the sequence of the GRanges.
}
\description{
Core function
}
