% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{binWrapper}
\alias{binWrapper}
\title{Internal function}
\usage{
binWrapper(grange, genome, MOTIF, BIN.WINDOW.SIZE, BIN.WINDOW.SLIDING,
  RANGE.FOR.SPECTRUM, FREQ)
}
\arguments{
\item{grange}{A GRanges object of length 1}

\item{genome}{A DNAStringSet object. Ideally, the sequence of an entire genome, 
obtained for instance by running 
\code{Biostrings::getSeq(BSgenome.Celegans.UCSC.ce11::BSgenome.Celegans.UCSC.ce11)}.}

\item{MOTIF}{String Oligonucleotide of interest.}

\item{BIN.WINDOW.SIZE}{Integer The width of the bins to split each primary bin
in (default 60).}

\item{BIN.WINDOW.SLIDING}{Integer The increment between secondary bins
(default 5).}

\item{RANGE.FOR.SPECTRUM}{Numeric vector The distances between nucleotides
to take into consideration when performing Fast Fourier Transform (default 1:50).}

\item{FREQ}{Float The frequence of the dinucleotide to study (default 1/10).}
}
\value{
Float The estimated periodicity score of the input GRanges
}
\description{
Internal function
}
