% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{binWrapper}
\alias{binWrapper}
\title{Internal function}
\usage{
binWrapper(grange, genome, motif, window_sliding_size, window_sliding_bin,
  range_spectrum, freq)
}
\arguments{
\item{grange}{A GRanges object of length 1}

\item{genome}{A DNAStringSet object. Ideally, the sequence of an entire 
genome}

\item{motif}{String Oligonucleotide of interest.}

\item{window_sliding_size}{Integer The width of the bins to split each primary
bin in (default 60).}

\item{window_sliding_bin}{Integer The increment between secondary bins
(default 5).}

\item{range_spectrum}{Numeric vector The distances between nucleotides
to take into consideration when performing Fast Fourier Transform 
(default seq_len(50)).}

\item{freq}{Float The frequence of the dinucleotide to study (default 1/10).}
}
\value{
Float The estimated periodicity score of the input GRanges
}
\description{
Internal function
}
