% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{char2BSgenome}
\alias{char2BSgenome}
\title{A function to get the BSgenome from genome ID}
\usage{
char2BSgenome(ID)
}
\arguments{
\item{ID}{character}
}
\value{
BSgenome object
}
\description{
This function switches a UCSC genome ID to the actual BSgenome
object. Currently, sacCer3, ce11, dm6, danRer10, mm10 and hg38 are 
supported.
}
\examples{
char2BSgenome('ce11')
}
