% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{chunkWrapper}
\alias{chunkWrapper}
\title{Internal function}
\usage{
chunkWrapper(
  chunk,
  chunks,
  genome,
  motif,
  step_size,
  range_spectrum,
  freq,
  logfile
)
}
\arguments{
\item{chunk}{A GRanges from the output list obtained with getChunks}

\item{genome}{DNAStringSet, BSgenome or genome ID}

\item{motif}{String Oligonucleotide of interest.}

\item{step_size}{Integer The increment between 
bins over GRanges (default 2).}

\item{range_spectrum}{Numeric vector The distances between nucleotides
to take into consideration when performing Fast Fourier Transform 
(default seq_len(50)).}

\item{freq}{Float The frequence of the dinucleotide to study 
(default 1/10).}

\item{logfile}{character, name of log file}
}
\value{
GRanges a GRanges object with a score column containing 
estimated periodicity
}
\description{
Internal function
}
