% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{chunkWrapper}
\alias{chunkWrapper}
\title{Internal function}
\usage{
chunkWrapper(
  chunk,
  genome,
  motif,
  genome_sliding_sliding,
  window_sliding_size,
  window_sliding_bin,
  range_spectrum,
  freq
)
}
\arguments{
\item{chunk}{A GRanges from the output list obtained with getChunks}

\item{genome}{DNAStringSet, BSgenome or genome ID}

\item{motif}{String Oligonucleotide of interest.}

\item{genome_sliding_sliding}{Integer The increment between 
bins over GRanges (default 2).}

\item{window_sliding_size}{Integer The width of the bins to split 
each primary bin in (default 60).}

\item{window_sliding_bin}{Integer The increment between secondary bins
(default 5).}

\item{range_spectrum}{Numeric vector The distances between nucleotides
to take into consideration when performing Fast Fourier Transform 
(default seq_len(50)).}

\item{freq}{Float The frequence of the dinucleotide to study 
(default 1/10).}
}
\value{
GRanges a GRanges object with a score column containing 
estimated periodicity
}
\description{
Internal function
}
