% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges.R
\name{deconvolveBidirectionalPromoters}
\alias{deconvolveBidirectionalPromoters}
\title{Core function}
\usage{
deconvolveBidirectionalPromoters(granges)
}
\arguments{
\item{granges}{A GRanges object}
}
\value{
GRanges with only '+' and '-' strands. GRanges with '*' strand 
have been duplicated and split into forward and reverse strands.
}
\description{
Core function
}
