% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DNAStringSet2GRanges}
\alias{DNAStringSet2GRanges}
\title{A function to generate a GRanges from a DNAStringSet}
\usage{
DNAStringSet2GRanges(seqs)
}
\arguments{
\item{seqs}{A DNAStringSet}
}
\value{
A GRanges object
}
\description{
A function to generate a GRanges from a DNAStringSet
}
\examples{
\dontrun{
    seq <- BSgenome.Celegans.UCSC.ce11::BSgenome.Celegans.UCSC.ce11
    DNAStringSet2GRanges(seq)
}
}
