% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{generatePeriodicityTrack}
\alias{generatePeriodicityTrack}
\title{Core function to generate a track of periodicity strenght of a given 
dinucleotide at a given frequency, over a set of chosen GRanges.}
\usage{
generatePeriodicityTrack(genome = NULL, granges, motif = "WW",
  extension = 1000, genome_sliding_size = 100,
  genome_sliding_sliding = 2, window_sliding_size = 100,
  window_sliding_bin = 5, range_spectrum = seq_len(50), freq = 1/10,
  cores = 12, bw_file = NULL)
}
\arguments{
\item{genome}{DNAStringSet object obtained from
\code{Biostrings::getSeq(BSgenome.xxx)}.}

\item{granges}{GRanges object (with seqnames overlapping 
the names of genome).}

\item{motif}{String Oligonucleotide of interest.}

\item{extension}{Integer The width the GRanges are going to 
be extended to (default 1000).}

\item{genome_sliding_size}{Integer The width of the bins to split the GRanges
objects in (default 100).}

\item{genome_sliding_sliding}{Integer The increment between bins over GRanges
(default 2).}

\item{window_sliding_size}{Integer The width of the bins to split each primary 
bin in (default 60).}

\item{window_sliding_bin}{Integer The increment between secondary bins
(default 5).}

\item{range_spectrum}{Numeric vector The distances between nucleotides
to take into consideration when performing Fast Fourier Transform 
(default seq_len(50)).}

\item{freq}{Float The frequence of the dinucleotide to study (default 1/10).}

\item{cores}{Integer Split the workload over several processors 
(default 12).}

\item{bw_file}{String. The name of the output bigWig track}
}
\value{
NULL A bigWig track in the working directory.
}
\description{
Core function to generate a track of periodicity strenght of a given 
dinucleotide at a given frequency, over a set of chosen GRanges.
}
