% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges.R
\name{getCovMatrix}
\alias{getCovMatrix}
\title{Internal function}
\usage{
getCovMatrix(g, bw, norm = "none", verbose = FALSE)
}
\arguments{
\item{g}{GRanges to map a track onto}

\item{bw}{RleList, a track from rtraklayer::import(..., as = 'Rle')}

\item{norm}{character, should the signal be normalized 
('none', 'zscore' or 'log2')?}

\item{verbose}{Boolean, should the function be verbose?}
}
\value{
A square numerical matrix of signal values over the GRanges
}
\description{
Used to compute the stranded coverage of a RleList over a 
GRanges as a square matrix.
}
