% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPI.R
\name{getFPI}
\alias{getFPI}
\title{A function to compute fold power increase (FPI)}
\usage{
getFPI(x, ...)
}
\arguments{
\item{x}{DNAStringSet, sequences of interest}

\item{...}{Additional arguments}
}
\value{
Several metrics including FPI, observed PSD, etc...
}
\description{
This function takes a set of sequences and a k-mer of interest, 
along with a given period. It calculates the FPI, as introduced
by Pich et al., Cell 2018.
}
\details{
The FPI is calculated as the ratio of the observed k-mer PSD 
at the chosen perdiod subtracted by the median k-mer PSD
in shuffled sequences, divided by the median k-mer PSD 
in shuffled sequences. The number of shufflings is specified 
by the n_shuffling argument.
}
\examples{
data(ce11_proms_seqs)
fpi <- getFPI(
    ce11_proms_seqs[1:10], 
    genome = 'ce11', 
    motif = 'TT', 
    BPPARAM = BiocParallel::SnowParam(workers = 1)
)
fpi$FPI
fpi$observed_PSD
fpi$shuffled_PSD
plotFPI(fpi)
}
