% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPI.R
\name{getFPI}
\alias{getFPI}
\title{A function to compute the fold power increase, as introduced by 
  Pich et al., Cell 2018.}
\usage{
getFPI(seqs, motif, period = 10, n_shuffle = 10,
  parallel_shuffling = 10, verbose = 1, ...)
}
\arguments{
\item{seqs}{DNAStringSet sequences of interest}

\item{motif}{character k-mer of interest}

\item{period}{integer Period of interest}

\item{n_shuffle}{integer Number of shuffling}

\item{parallel_shuffling}{integer Number of threads to use to split
shuffling}

\item{verbose}{integer Should the function be verbose?}

\item{...}{Additional arguments}
}
\value{
list Several metrics included (FPI, observed PSD, etc...)
}
\description{
A function to compute the fold power increase, as introduced by 
  Pich et al., Cell 2018.
}
