% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicity.R
\name{getPeriodicity.DNAString}
\alias{getPeriodicity.DNAString}
\title{Core function - NOT WORKING. DO NOT USE}
\usage{
\method{getPeriodicity}{DNAString}(seq, motif = "TT",
  subseq_len = NULL, n_occurences = 200, ...)
}
\arguments{
\item{seq}{a DNAString}

\item{motif}{a dinucleotide of interest}

\item{subseq_len}{The length of sub-sequences}

\item{n_occurences}{The targeted number of pairs of dinucleotides to obtain.}
}
\value{
List NULL
}
\description{
Core function - NOT WORKING. DO NOT USE
}
