% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicity.R
\name{getPeriodicity.DNAString}
\alias{getPeriodicity.DNAString}
\title{A function to compute k-mer periodicity in a sequence.}
\usage{
\method{getPeriodicity}{DNAString}(x, ...)
}
\arguments{
\item{x}{a DNAString}

\item{...}{other parameters forwarded to getPeriodicity.DNAStringSet()}
}
\value{
A list containing the results of getPeriodicity function.  
\itemize{
    \item The dists vector is the raw vector of all distances between 
    any possible dinucleotide. 
    \item The hist data.frame is the distribution of distances
    over range_spectrum. 
    \item The normalized_hist is the raw hist, 
    normalized for decay over increasing distances. 
    \item The spectra object is the output of 
    the FFT applied over normalized_hist. 
    \item The PSD data frame is the power 
    spectral density scores over given  frequencies. 
    \item The motif object is the dinucleotide being analysed.
}
}
\description{
This function takes a single sequence and a k-mer of interest, 
map a k-mer of interest in the sequence, computes all the 
pairwise distances (distogram), normalize it for distance decay, 
and computes the resulting power spectral density of the 
normalized distogram.
}
\examples{
data(ce11_proms_seqs)
periodicity_result <- getPeriodicity(
    ce11_proms_seqs[[8]],
    motif = 'TT', 
    cores = 1
)
head(periodicity_result$PSD)
plotPeriodicityResults(periodicity_result)
}
