% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicity.R
\name{getPeriodicity.DNAStringSet}
\alias{getPeriodicity.DNAStringSet}
\title{Core function}
\usage{
\method{getPeriodicity}{DNAStringSet}(seqs, motif = "WW",
  RANGE_FOR_SPECTRUM = 1:100, period = seq(2, 20, 1), plot = FALSE,
  cores = 2, verbose = TRUE)
}
\arguments{
\item{seqs}{a DNAStringSet}

\item{motif}{a dinucleotide of interest}

\item{RANGE_FOR_SPECTRUM}{Numeric vector The distances between nucleotides
to take into consideration when performing Fast Fourier Transform (default 1:50).}

\item{period}{Vector a numerical vector of periods to extract.}

\item{plot}{Boolean Should the FFT results be plotted?}

\item{cores}{Integer How many processors should be used to split to work?}

\item{verbose}{Boolean}
}
\value{
List a list containing the results of getPeriodicity function. The 
dists vector is the raw vector of all distances between any possible dinucleotide. 
The hist data.frame is the distribution of distances over RANGE_FOR_SPECTRUM. 
The normalized_hist is the raw hist, normalized for decay over increasing distances.
The spectra object is the output of the FFT applied over normalized_hist. 
The PSD data frame is the power spectrum density scores over given frequencies.
The signal_to_noise_ratio is a data.frame containing enrichment scores of TT
periodicity, for the periods in the period vector. The motif object is the 
dinucleotide being analysed.
}
\description{
Core function
}
