% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicity.R
\name{getPeriodicity.DNAStringSet}
\alias{getPeriodicity.DNAStringSet}
\title{A function to compute k-mer periodicity in sequence(s).}
\usage{
\method{getPeriodicity}{DNAStringSet}(
  x,
  motif = "WW",
  range_spectrum = seq(1, 200),
  cores = 1,
  roll = 3,
  verbose = TRUE,
  sample = 0,
  doZscore = FALSE,
  skip_shuffling = TRUE,
  ...
)
}
\arguments{
\item{x}{a DNAStringSet}

\item{motif}{a dinucleotide of interest}

\item{range_spectrum}{Numeric vector The distances between nucleotides
to take into consideration when performing Fast Fourier Transform.}

\item{cores}{Integer How many processors should be used to parallelize
the mapping?}

\item{roll}{Integer Window to smooth the distribution of pairwise distances
(default: 3, to discard the 3-bp periodicity of dinucleotides which 
can be very strong in vertebrate genomes)}

\item{verbose}{Boolean}

\item{sample}{Integer if > 0, will randomly sample this many integers
from the dists vector before normalization. This ensures consistency 
when looking at periodicity in different genomes, since different
genomes will have different GC percent}

\item{doZscore}{Boolean should the normalized dampened signal be z-scored?}

\item{skip_shuffling}{Boolean should the shuffling sequences be done?}

\item{...}{additional parameters}
}
\value{
A list containing the results of getPeriodicity function.  
\itemize{
    \item The dists vector is the raw vector of all distances between 
    any possible dinucleotide. 
    \item The hist data.frame is the distribution of distances
    over range_spectrum. 
    \item The normalized_hist is the raw hist, 
    normalized for decay over increasing distances. 
    \item The spectra object is the output of 
    the FFT applied over normalized_hist. 
    \item The PSD data frame is the power 
    spectral density scores over given  frequencies. 
    \item The motif object is the dinucleotide being analysed.
}
}
\description{
This function takes a set of sequences and a k-mer of interest, 
map a k-mer of interest in these sequences, computes all the 
pairwise distances (distogram), normalize it for distance decay, 
and computes the resulting power spectral density of the 
normalized distogram.
}
\examples{
data(ce11_proms_seqs)
periodicity_result <- getPeriodicity(
    ce11_proms_seqs[1:10],
    motif = 'TT', 
    cores = 1
)
head(periodicity_result$PSD)
plotPeriodicityResults(periodicity_result)
}
