% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicity.R
\name{getPeriodicity.GRanges}
\alias{getPeriodicity.GRanges}
\title{Core function}
\usage{
\method{getPeriodicity}{GRanges}(x, genome = "ce11", ...)
}
\arguments{
\item{x}{a GRanges}

\item{genome}{DNAStringSet object. The sequence of an entire genome, 
obtained for instance by running 
\code{Biostrings::getSeq(
    BSgenome.Celegans.UCSC.ce11::BSgenome.Celegans.UCSC.ce11
)}.}

\item{...}{other parameters forwarded to getPeriodicity.DNAStringSet()}
}
\value{
List a list containing the results of getPeriodicity function. 
    The dists vector is the raw vector of all distances between any 
    possible dinucleotide. The hist data.frame is the distribution of 
    distances over RANGE_FOR_SPECTRUM. The normalized_hist is the raw 
    hist, normalized for  decay over increasing distances. The spectra 
    object is the output of the FFT applied over normalized_hist. 
    The PSD data frame is the power spectral density scores over given 
    frequencies. The motif object is the dinucleotide being analysed.
}
\description{
Core function
}
