% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicity.R
\name{getPeriodicity.GRanges}
\alias{getPeriodicity.GRanges}
\title{Core function}
\usage{
\method{getPeriodicity}{GRanges}(granges,
  genome = Biostrings::getSeq(BSgenome.Celegans.UCSC.ce11::BSgenome.Celegans.UCSC.ce11),
  ...)
}
\arguments{
\item{granges}{a GRanges}

\item{genome}{DNAStringSet object. Ideally, the sequence of an entire genome, 
obtained for instance by running 
\code{Biostrings::getSeq(BSgenome.Celegans.UCSC.ce11::BSgenome.Celegans.UCSC.ce11)}.}

\item{...}{other parameters required in getPeriodicity}
}
\value{
List a list containing the results of getPeriodicity function. The 
dists vector is the raw vector of all distances between any possible dinucleotide. 
The hist data.frame is the distribution of distances over RANGE_FOR_SPECTRUM. 
The normalized_hist is the raw hist, normalized for decay over increasing distances.
The spectra object is the output of the FFT applied over normalized_hist. 
The PSD data frame is the power spectrum density scores over given frequencies.
The signal_to_noise_ratio is a data.frame containing enrichment scores of TT
periodicity, for the periods in the period vector. The motif object is the 
dinucleotide being analysed.
}
\description{
Core function
}
