% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicity.R
\name{getPeriodicity}
\alias{getPeriodicity}
\title{A function to compute k-mer periodicity in sequence(s).}
\usage{
getPeriodicity(x, ...)
}
\arguments{
\item{x}{a DNAStringSet, or a GRanges}

\item{...}{additional parameters}
}
\value{
A list containing the results of getPeriodicity function.  
\itemize{
    \item The dists vector is the raw vector of all distances between 
    any possible dinucleotide. 
    \item The hist data.frame is the distribution of distances
    over range_spectrum. 
    \item The normalized_hist is the raw hist, 
    normalized for decay over increasing distances. 
    \item The spectra object is the output of 
    the FFT applied over normalized_hist. 
    \item The PSD data frame is the power 
    spectral density scores over given  frequencies. 
    \item The motif object is the dinucleotide being analysed.
}
}
\description{
This function takes a set of sequences and a k-mer of interest, 
map a k-mer of interest in these sequences, computes all the 
pairwise distances (distogram), normalize it for distance decay, 
and computes the resulting power spectral density of the 
normalized distogram.
}
\examples{
data(ce11_proms_seqs)
periodicity_result <- getPeriodicity(
    ce11_proms_seqs[1:100],
    motif = 'TT'
)
plotPeriodicityResults(periodicity_result)
#
data(ce11_proms)
periodicity_result <- getPeriodicity(
    ce11_proms[1:100],
    genome = 'ce11',
    motif = 'TT'
)
head(periodicity_result$PSD)
plotPeriodicityResults(periodicity_result)
}
