% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPI.R
\name{getSignificantPeriods}
\alias{getSignificantPeriods}
\title{A function used to compute FPIs and associated p-values}
\usage{
getSignificantPeriods(fpi)
}
\arguments{
\item{fpi}{result of getFPI() function or getPeriodicity() function with
n_shuffling > 1}
}
\value{
A table showing FPI and p-value for each observed PSD value
}
\description{
A function used to compute FPIs and associated p-values
}
\examples{
data(ce11_TSSs)
fpi <- getFPI(
    ce11_TSSs[['Ubiq.']][1:10], 
    genome = 'ce11', 
    motif = 'TT', 
    cores_shuffling = 1
)
head(getSignificantPeriods(fpi))
}
