% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na.remove}
\alias{na.remove}
\title{A function to remove NAs from a vector}
\usage{
na.remove(x)
}
\arguments{
\item{x}{vector}
}
\value{
A vector without NAs
}
\description{
This function is 'tidyverse-friendly', i.e. it takes a vector and
returns the same vector minus the NA values
}
