% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na.replace}
\alias{na.replace}
\title{A tidyverse-compatible function to replace NAs in a vector by a given value}
\usage{
na.replace(x, value)
}
\arguments{
\item{x}{vector}

\item{value}{Replace NAs by this variable}
}
\value{
A vector with NA replaced by value
}
\description{
A tidyverse-compatible function to replace NAs in a vector by a given value
}
