% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{namedListToLongFormat}
\alias{namedListToLongFormat}
\title{A function to easily coerce a named list into a long data.frame}
\usage{
namedListToLongFormat(x)
}
\arguments{
\item{x}{A named list of vectors or data.frames}
}
\value{
A long data frame
}
\description{
This function takes a named list of data.frames. For each data.frame, 
the name of the data.frame is added in a new column, and the resulting
data.frames are all bound together using rbind.
to
}
