% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicity.R
\name{normalizeHistogram}
\alias{normalizeHistogram}
\title{Internal function to normalize a pairwise distance}
\usage{
normalizeHistogram(hist, roll = 1, doZscore = TRUE,
  roll_smoothed.h = 10)
}
\arguments{
\item{hist}{Vector a numeric vector}
}
\value{
a normalized vector
}
\description{
Internal function to normalize a pairwise distance
}
