% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicity.R
\name{normalizeHistogram}
\alias{normalizeHistogram}
\title{Internal function}
\usage{
normalizeHistogram(hist, roll = 1, doZscore = TRUE, roll_smoothed.h = 10)
}
\arguments{
\item{hist}{Vector a numeric vector}

\item{roll}{Integer window used to roll the flatten histogram}

\item{doZscore}{Boolean should the normalized dampened signal be z-scored?}

\item{roll_smoothed.h}{Integer window used to flatten the histogram}
}
\value{
a normalized vector
}
\description{
This function normalize a distogram for distance decay
}
