% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{partitionBin}
\alias{partitionBin}
\title{Internal function}
\usage{
partitionBin(grange, genome, BIN.WINDOW.SIZE, BIN.WINDOW.SLIDING)
}
\arguments{
\item{grange}{A GRanges object of length 1}

\item{genome}{A DNAStringSet object. Ideally, the sequence of an entire 
genome}

\item{BIN.WINDOW.SIZE}{Integer The width of the bins to split each primary
bin in (default 60).}

\item{BIN.WINDOW.SLIDING}{Integer The increment between secondary bins
(default 5).}
}
\value{
list A GRanges object
}
\description{
Internal function
}
