% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{partitionGenome}
\alias{partitionGenome}
\title{Internal function}
\usage{
partitionGenome(
  genome,
  granges.extended_large,
  granges.extended,
  genome_sliding_size,
  genome_sliding_sliding
)
}
\arguments{
\item{genome}{DNAStringSet, BSgenome or genome ID}

\item{granges.extended_large}{A GRanges object.}

\item{granges.extended}{A GRanges object.}

\item{genome_sliding_size}{An integer. The width of the bins to split 
the GRanges objects in (default 100).}

\item{genome_sliding_sliding}{An integer. The increment between bins 
over GRanges (default 2).}
}
\value{
granges.partionned A Granges object corresponding to the binned 
genome.
}
\description{
Internal function
}
