% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges.R
\name{plotAggregateCoverage.CompressedRleList}
\alias{plotAggregateCoverage.CompressedRleList}
\title{A function to plot aggregated signals (from a CompressedRleList) 
  over sets of GRanges}
\usage{
\method{plotAggregateCoverage}{CompressedRleList}(bw, granges, ...)
}
\arguments{
\item{bw}{a single signal track (CompressedRleList class)}

\item{granges}{a GRanges object or a list of GRanges}
}
\value{
A plot of aggregated signals
}
\description{
A function to plot aggregated signals (from a CompressedRleList) 
  over sets of GRanges
}
