% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges.R
\name{plotAggregateCoverage.CompressedRleList}
\alias{plotAggregateCoverage.CompressedRleList}
\title{Core function}
\usage{
\method{plotAggregateCoverage}{CompressedRleList}(bw, granges, ...)
}
\arguments{
\item{bw}{a single signal track CompressedRleList class}
}
\value{
A plot
}
\description{
Core function
}
