% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges.R
\name{plotAggregateCoverage.list}
\alias{plotAggregateCoverage.list}
\title{A function to plot aggregated signals over sets of GRanges}
\usage{
\method{plotAggregateCoverage}{list}(
  x,
  granges,
  colors = NULL,
  xlab = "Center of elements",
  ylab = "Score",
  xlim = NULL,
  ylim = NULL,
  quartiles = c(0.025, 0.975),
  verbose = FALSE,
  bin = 1,
  plot_central = TRUE,
  split_by_granges = TRUE,
  split_by_track = FALSE,
  free_scales = FALSE,
  run_in_parallel = TRUE,
  norm = "none",
  ...
)
}
\arguments{
\item{x}{several signal tracks (SimpleRleList or CompressedRleList class) 
grouped in a named list}

\item{granges}{a GRanges object or a list of GRanges}

\item{colors}{a vector of colors}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{xlim}{y axis limits}

\item{ylim}{y axis limits}

\item{quartiles}{Which quantiles to use to determine y scale automatically?}

\item{verbose}{Boolean}

\item{bin}{Integer Width of the window to use to smooth values}

\item{plot_central}{Boolean Draw a vertical line at 0}

\item{split_by_granges}{Boolean Facet plots by the sets of GRanges}

\item{split_by_track}{Boolean Facet plots by the sets of signal tracks}

\item{free_scales}{Boolean Should each facet have independent y-axis scales?}

\item{run_in_parallel}{Boolean Should the plots be computed in parallel using 
mclapply?}

\item{norm}{character Should the signals be normalized 
('none', 'zscore' or 'log2')?}

\item{...}{additional parameters}
}
\value{
A plot of aggregated signals
}
\description{
This function takes one or several RleList genomic tracks (imported
by rtraklayer::import(..., as = 'Rle')) and one or several GRanges 
objects. It computes coverage of the GRanges by the genomic tracks
and returns an aggregate coverage plot.
}
\examples{
library(GenomicRanges)
data(ce11_ATACseq)
data(ce11_WW_10bp)
data(ce11_proms)
proms <- resize(ce11_proms[1:100], fix = 'center', width = 400)
p1 <- plotAggregateCoverage(
    list(
        'atac' = ce11_ATACseq, 
        'WW_10bp' = ce11_WW_10bp
    ), 
    proms,
    norm = 'zscore'
)
p1
p2 <- plotAggregateCoverage(
    list(
        'ATAC-seq' = ce11_ATACseq, 
        'WW 10-bp periodicity' = ce11_WW_10bp
    ), 
    list(
        'Ubiq & Germline promoters' = 
            proms[proms$which.tissues \%in\% c('Ubiq.', 'Germline')],
        'Other promoters' = 
            proms[!(proms$which.tissues \%in\% c('Ubiq.', 'Germline'))]
    ), 
    norm = 'zscore'
)
p2
p3 <- plotAggregateCoverage(
    list(
        'ATAC-seq' = ce11_ATACseq, 
        'WW 10-bp periodicity' = ce11_WW_10bp
    ), 
    list(
        'Ubiq & Germline promoters' = 
            proms[proms$which.tissues \%in\% c('Ubiq.', 'Germline')],
        'Other promoters' = 
            proms[!(proms$which.tissues \%in\% c('Ubiq.', 'Germline'))]
), 
split_by_granges = FALSE,
split_by_track = TRUE,
norm = 'zscore'
)
p3
}
