% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges.R
\name{plotAggregateCoverage}
\alias{plotAggregateCoverage}
\title{A function to plot aggregated signals over sets of GRanges}
\usage{
plotAggregateCoverage(x, ...)
}
\arguments{
\item{x}{CompressedRleList or SimpleRleList (in lists or single)}

\item{...}{additional parameters}
}
\value{
A plot of aggregated signals
}
\description{
A function to plot aggregated signals over sets of GRanges
}
