% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges.R
\name{plotAggregateCoverage}
\alias{plotAggregateCoverage}
\title{Core function}
\usage{
plotAggregateCoverage(x, ...)
}
\arguments{
\item{x}{a single signal track (in SimpleRleList or CompressedRleList class) 
or a set of signal tracks (in list class)}
}
\value{
A plot
}
\description{
Core function
}
