% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotFPI}
\alias{plotFPI}
\title{Plot the output of getFPI()}
\usage{
plotFPI(fpi, periods = c(2, 20), fdr_threshold = 0.01)
}
\arguments{
\item{fpi}{The output of getFPI function.}

\item{periods}{Vector a numerical vector of length 2, to specify the
x-axis limits}

\item{fdr_threshold}{Float, significance threshold}
}
\value{
ggplot A ggplot
}
\description{
This function plots some results from the result of getFPI(). 
It plots the observed PSD values in red and all the shuffled 
runs in grey.
}
\examples{
data(ce11_proms_seqs)
fpi <- getFPI(
    ce11_proms_seqs[1:10], 
    motif = 'TT', 
    cores_computing = 1
)
plotFPI(fpi)
}
