% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotFPI}
\alias{plotFPI}
\title{Plotting function}
\usage{
plotFPI(fpi, periods = c(2, 20), s = 0.05)
}
\arguments{
\item{fpi}{The output of getFPI function.}

\item{periods}{Vector a numerical vector of length 2, to specify the
x-axis limits}

\item{s}{float span parameter for loess smooth}
}
\value{
ggplot A ggplot
}
\description{
Plotting function
}
