% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPeriodicityResults}
\alias{plotPeriodicityResults}
\title{Plotting function}
\usage{
plotPeriodicityResults(
  results,
  periods = c(2, 20),
  filter_periods = TRUE,
  skip_shuffling = FALSE,
  facet_control = TRUE,
  xlim = NULL,
  ...
)
}
\arguments{
\item{results}{The output of getPeriodicity function.}

\item{periods}{Vector a numerical vector of length 2, to specify the
x-axis limits}

\item{filter_periods}{Boolean Should the x-axis be 
constrained to the periods?}

\item{skip_shuffling}{Boolean should the shuffling sequences be done?}

\item{facet_control}{Boolean should the shuffling plots be faceted?}

\item{xlim}{Integer x axis upper limit in raw and norm. distograms}

\item{...}{Additional theme arguments passed to theme_ggplot2()}
}
\value{
list A list containing four ggplots
}
\description{
Plotting function
}
