% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPeriodicityResults}
\alias{plotPeriodicityResults}
\title{Plotting function}
\usage{
plotPeriodicityResults(results, HIST_YLIM = c(-1.5, 1.5))
}
\arguments{
\item{results}{The output of getPeriodicity function.}

\item{HIST_YLIM}{Vector a numerical vector of length 2, to specify the y-axis 
limits of the first plot (normalized distribution of pairwise distances).}
}
\value{
list A list containing three ggplots
}
\description{
Plotting function
}
