% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sampleGRanges.BSgenome}
\alias{sampleGRanges.BSgenome}
\title{A function to sample GRanges from BSgenome}
\usage{
\method{sampleGRanges}{BSgenome}(x, ...)
}
\arguments{
\item{x}{a BSgenome object}

\item{...}{Additional parameters}
}
\value{
A GRanges object of length n
}
\description{
This function takes a BSgenome and returns a GRanges object. 
This function is useful to sample sequences from a full genome.
}
\examples{
g <- sampleGRanges(
    BSgenome.Scerevisiae.UCSC.sacCer3::BSgenome.Scerevisiae.UCSC.sacCer3, 
    w = 150, n = 100
)
g
g$seq
}
