% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sampleGRanges.character}
\alias{sampleGRanges.character}
\alias{sampleGenome}
\title{A function to sample GRanges from genome IDs}
\usage{
\method{sampleGRanges}{character}(x, ...)

sampleGenome(x, ...)
}
\arguments{
\item{x}{a genome ID ()}

\item{...}{Additional parameters}
}
\value{
A GRanges object of length n
}
\description{
This function takes a BSgenome UCSC ID and returns a GRanges object. 
This function is useful to sample sequences from a full genome. 
Currently, sacCer3, ce11, dm6, danRer10, mm10 and hg38 are 
supported.

This function is an alias for the sampleGRanges.character method.
It takes a BSgenome ID and returns a GRanges object. 
This function is useful to sample sequences from a full genome. 
Currently, sacCer3, ce11, dm6, danRer10, mm10 and hg38 are 
supported.
}
\examples{
sampleGRanges('sacCer3', width = 150, n = 100)
sampleGenome('sacCer3', w = 150, n = 100)
}
