% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sampleGRanges.DNAStringSet}
\alias{sampleGRanges.DNAStringSet}
\title{A function to sample GRanges within DNAStringSet}
\usage{
\method{sampleGRanges}{DNAStringSet}(x, ...)
}
\arguments{
\item{x}{a DNAStringSet object}

\item{...}{Additional parameters}
}
\value{
A GRanges object of length n
}
\description{
This function takes a DNAStringSet (for instance a set of sequences
obtained from a BSgenome object) and returns a GRanges object. 
The ranges are contained within the DNAStringSet object. This function
is useful to sample sequences from a full genome.
}
\examples{
library(Biostrings)
seqs <- getSeq(BSgenome.Scerevisiae.UCSC.sacCer3::
    BSgenome.Scerevisiae.UCSC.sacCer3)
sampleGRanges(seqs, n = 100, w = 100)
}
