% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigwigs.R
\name{scaleBigWigs}
\alias{scaleBigWigs}
\title{A function to z-score a bigWig track imported by rtracklayer.}
\usage{
scaleBigWigs(bigWigs)
}
\arguments{
\item{bigWigs}{a RleList}
}
\value{
a list of scaled Rle (z-score)
}
\description{
This function takes a RleList (usually obtained by 
rtracklayer::import(..., as = 'Rle')) and scales each chromosome
idependently by Z-score.
}
\examples{
data(ce11_WW_10bp)
scaleBigWigs(ce11_WW_10bp)
}
