% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shuffleSeq}
\alias{shuffleSeq}
\title{A function to shuffle sequence(s)}
\usage{
shuffleSeq(dna, order = 1)
}
\arguments{
\item{dna}{DNAString or DNAStringSet}

\item{order}{Integer, which order to take into consideration for shuffling
(requires Python ushuffle library)}
}
\value{
A DNAString or DNAStringSet
}
\description{
This function takes a DNAString or DNAStringSet object and simply 
shuffles the order individual nucleotides.
}
\examples{
shuffleSeq('ACGTGGGCTATTAGCTACTGTACGTG')
}
