% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_ggplot2}
\alias{theme_ggplot2}
\title{Own ggplot2 theming function}
\usage{
theme_ggplot2(
  grid = TRUE,
  border = TRUE,
  base_family = NULL,
  base_size = 8,
  plot_title_family = base_family,
  plot_title_size = 12,
  plot_title_face = "plain",
  plot_title_margin = 5,
  subtitle_size = 11,
  subtitle_face = "plain",
  subtitle_margin = 5,
  strip_text_family = base_family,
  strip_text_size = 10,
  strip_text_face = "bold",
  caption_size = 9,
  caption_face = "plain",
  caption_margin = 3,
  axis_text_size = base_size,
  axis_title_family = base_family,
  axis_title_size = 9,
  axis_title_face = "plain",
  axis_title_just = "rt",
  panel_spacing = grid::unit(2, "lines"),
  grid_col = "#cccccc",
  plot_margin = margin(12, 12, 12, 12),
  axis_col = "#cccccc",
  axis = FALSE,
  ticks = FALSE
)
}
\value{
theme A ggplot theme
}
\description{
Own ggplot2 theming function
}
