% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges.R
\name{withSeq}
\alias{withSeq}
\title{Tidyverse-compatible function to add sequence to an existing GRanges}
\usage{
withSeq(granges, genome)
}
\arguments{
\item{granges}{A GRanges object}

\item{genome}{DNAStringSet object. Ideally, the sequence of an entire genome, 
obtained for instance by running}
}
\value{
GRanges with a seq column containing the sequence of the GRanges.
}
\description{
Tidyverse-compatible function to add sequence to an existing GRanges
}
